+++ db/migration/000011_create_resources_table.up.sql
@@ -1,8 +1,8 @@
 CREATE TABLE resources (
     "id" BIGSERIAL PRIMARY KEY,
     "resource_name" varchar NOT NULL UNIQUE,
-    "current" FLOAT DEFAULT (0) NOT NULL ,
-    "availed" FLOAT NOT NULL,
+    "current" NUMERIC DEFAULT (0) NOT NULL ,
+    "availed" NUMERIC NOT NULL,
     "created_at" timestamptz NOT NULL DEFAULT (now()),
     "created_by" VARCHAR NOT NULL,
     "updated_on" timestamptz NOT NULL DEFAULT (now()),
